chore: SearchMap 제거 및 Search로 이관
