+++ client/Targets/Presentation/Search/Implementations/Sources/Search/Views/SearchMapMarkerAdapter.swift
